refactor: split config loader
