+++ tf_pwa/config_loader/__init__.py
@@ -2,9 +2,8 @@ from .config_loader import (
     ConfigLoader,
     FitResult,
     PlotParams,
-    export_legend,
-    hist_error,
-    hist_line,
     validate_file_name,
 )
+from .extra import *
 from .multi_config import MultiConfig
+from .plot import export_legend, hist_error, hist_line
